+++ src/Constants/ApiResponseCodes.php
@@ -66,6 +66,7 @@ class ApiResponseCodes
     const API_ERROR_AMOUNT_IS_MISSING                           = 'API.340.200.130';
     const API_ERROR_CUSTOMER_DOES_NOT_EXIST                     = 'API.410.100.100';
     const API_ERROR_CUSTOMER_ID_ALREADY_EXISTS                  = 'API.410.200.010';
+    const API_ERROR_ADDRESS_NAME_TO_LONG                        = 'API.410.200.031';
     const API_ERROR_CUSTOMER_CAN_NOT_BE_FOUND                   = 'API.500.100.100';
     const API_ERROR_REQUEST_DATA_IS_INVALID                     = 'API.500.300.999';
     const API_ERROR_RECURRING_PAYMENT_NOT_SUPPORTED             = 'API.500.550.004';