+++ CHANGELOG.md
@@ -3,6 +3,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.7.0][1.2.7.0]
+
+### Added
+*   Http headers to debug log.
+*   PID to debug log.
+*   PHP7.4 support.
+*   Handles for automatic testing to examples.
+
+### Changed
+*   Update example implementation for Hire Purchase direct debit payment type (FlexiPay® Rate).
+*   Write log messages from examples to examples\log\example.log instead of error.log.
+*   Several minor changes.
+
+### Fix
+*   Ensure locale strings comply to standard.
+
 ## [1.2.6.0][1.2.6.0]
 
 ### Added
@@ -364,3 +380,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.2.5.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.4.0..1.2.5.0
 [1.2.5.1]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.5.0..1.2.5.1
 [1.2.6.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.5.1..1.2.6.0
+[1.2.7.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.6.0..1.2.7.0